Added the last queries in the mongoDB implementation. Added unit-tests to the newly implemented methods. Changed the interface to support doubles instead of floats on the coordinates.

+++ src/main/java/data/IQuery.java
@@ -31,7 +31,7 @@ public interface IQuery {
      * @param distance radius of the search distance
      * @return list of BookDTO's
      */
-    List<BookDTO> getBooksByGeoLocation(float lon, float lat, float distance);
+    List<BookDTO> getBooksByGeoLocation(double lon, double lat, double distance);
 
     /**
      * Given a booktitle, the method returns a list of cities mentioned in the book
